rca-backend/requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangGraph and AI
langgraph==0.2.45
langchain==0.3.9
langchain-core==0.3.18
langchain-anthropic==0.3.0
langchain-openai==0.2.10
anthropic==0.40.0
openai==1.59.3

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# Database clients
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
boto3==1.34.19

# Data processing
pandas==2.1.4
numpy==1.26.3

# Slack and Atlassian
slack-sdk==3.26.2
atlassian-python-api==3.41.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
